requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
numpy>=1.24.0
mysql-connector-python>=9.0.0
flask>=3.0.0
flask-cors>=4.0.0

